Removed Manual Config in ChildProfileActivity and ChwRepository
