+++ opensrp-chw/src/main/java/org/smartregister/chw/repository/ChwRepository.java
@@ -9,10 +9,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.smartregister.AllConstants;
 import org.smartregister.chw.BuildConfig;
 import org.smartregister.chw.application.ChwApplication;
-import org.smartregister.chw.util.Country;
-import org.smartregister.chw.util.RepositoryUtils;
 import org.smartregister.configurableviews.repository.ConfigurableViewsRepository;
-import org.smartregister.domain.db.Column;
 import org.smartregister.immunization.ImmunizationLibrary;
 import org.smartregister.immunization.repository.RecurringServiceRecordRepository;
 import org.smartregister.immunization.repository.RecurringServiceTypeRepository;
@@ -20,7 +17,6 @@ import org.smartregister.immunization.repository.VaccineNameRepository;
 import org.smartregister.immunization.repository.VaccineRepository;
 import org.smartregister.immunization.repository.VaccineTypeRepository;
 import org.smartregister.immunization.util.IMDatabaseUtils;
-import org.smartregister.repository.AlertRepository;
 import org.smartregister.repository.EventClientRepository;
 import org.smartregister.repository.Repository;
 import org.smartregister.repository.SettingsRepository;
@@ -93,32 +89,8 @@ public class ChwRepository extends Repository {
         Log.w(ChwRepository.class.getName(),
                 "Upgrading database from version " + oldVersion + " to "
                         + newVersion + ", which will destroy all old data");
-        int upgradeTo = oldVersion + 1;
-        while (upgradeTo <= newVersion) {
-            switch (upgradeTo) {
-                case 2:
-                    upgradeToVersion2(db);
-                    break;
-                case 3:
-                    upgradeToVersion3(db);
-                    break;
-                case 4:
-                    upgradeToVersion4(db);
-                    break;
-                case 5:
-                    upgradeToVersion5(db);
-                    break;
-                case 6:
-                    upgradeToVersion6(db);
-                    break;
-                case 7:
-                    upgradeToVersion7(db);
-                    break;
-                default:
-                    break;
-            }
-            upgradeTo++;
-        }
+
+        ChwRepositoryFlv.onUpgrade(context, db, oldVersion, newVersion);
     }
 
 
@@ -193,92 +165,4 @@ public class ChwRepository extends Repository {
         }
     }
     */
-
-    private void upgradeToVersion2(SQLiteDatabase db) {
-        try {
-            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_EVENT_ID_COL);
-            db.execSQL(VaccineRepository.EVENT_ID_INDEX);
-            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_FORMSUBMISSION_ID_COL);
-            db.execSQL(VaccineRepository.FORMSUBMISSION_INDEX);
-
-            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_OUT_OF_AREA_COL);
-            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_OUT_OF_AREA_COL_INDEX);
-
-//            EventClientRepository.createTable(db, EventClientRepository.Table.path_reports, EventClientRepository.report_column.values());
-            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_HIA2_STATUS_COL);
-
-            IMDatabaseUtils.accessAssetsAndFillDataBaseForVaccineTypes(context, db);
-
-        } catch (Exception e) {
-            Log.e(TAG, "upgradeToVersion2 " + Log.getStackTraceString(e));
-        }
-
-    }
-
-    private void upgradeToVersion3(SQLiteDatabase db) {
-        try {
-            Column[] columns = {EventClientRepository.event_column.formSubmissionId};
-            EventClientRepository.createIndex(db, EventClientRepository.Table.event, columns);
-
-            db.execSQL(VaccineRepository.ALTER_ADD_CREATED_AT_COLUMN);
-            VaccineRepository.migrateCreatedAt(db);
-
-            db.execSQL(RecurringServiceRecordRepository.ALTER_ADD_CREATED_AT_COLUMN);
-            RecurringServiceRecordRepository.migrateCreatedAt(db);
-        } catch (Exception e) {
-            Log.e(TAG, "upgradeToVersion3 " + Log.getStackTraceString(e));
-        }
-        try {
-            Column[] columns = {EventClientRepository.event_column.formSubmissionId};
-            EventClientRepository.createIndex(db, EventClientRepository.Table.event, columns);
-
-
-        } catch (Exception e) {
-            Log.e(TAG, "upgradeToVersion3 " + e.getMessage());
-        }
-    }
-
-    private void upgradeToVersion4(SQLiteDatabase db) {
-        try {
-            db.execSQL(AlertRepository.ALTER_ADD_OFFLINE_COLUMN);
-            db.execSQL(AlertRepository.OFFLINE_INDEX);
-            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_TEAM_COL);
-            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_TEAM_ID_COL);
-            db.execSQL(RecurringServiceRecordRepository.UPDATE_TABLE_ADD_TEAM_COL);
-            db.execSQL(RecurringServiceRecordRepository.UPDATE_TABLE_ADD_TEAM_ID_COL);
-        } catch (Exception e) {
-            Log.e(TAG, "upgradeToVersion4 " + Log.getStackTraceString(e));
-        }
-
-    }
-
-    private void upgradeToVersion5(SQLiteDatabase db) {
-        try {
-            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_CHILD_LOCATION_ID_COL);
-            db.execSQL(RecurringServiceRecordRepository.UPDATE_TABLE_ADD_CHILD_LOCATION_ID_COL);
-        } catch (Exception e) {
-            Log.e(TAG, "upgradeToVersion5 " + Log.getStackTraceString(e));
-        }
-    }
-
-    private void upgradeToVersion6(SQLiteDatabase db) {
-        try {
-            if (BuildConfig.BUILD_COUNTRY == Country.TANZANIA) {
-                for (String query : RepositoryUtils.UPGRADE_V6) {
-                    db.execSQL(query);
-                }
-            }
-        } catch (Exception e) {
-            Log.e(TAG, "upgradeToVersion6 " + Log.getStackTraceString(e));
-        }
-    }
-
-    private void upgradeToVersion7(SQLiteDatabase db) {
-        try {
-            db.execSQL(HomeVisitRepository.UPDATE_TABLE_ADD_VACCINE_NOT_GIVEN);
-            db.execSQL(HomeVisitRepository.UPDATE_TABLE_ADD_SERVICE_NOT_GIVEN);
-        } catch (Exception e) {
-            Log.e(TAG, "upgradeToVersion6 " + Log.getStackTraceString(e));
-        }
-    }
 }
