+++ opensrp-chw/src/wcaro/java/org/smartregister/chw/activity/ChildProfileActivityFlv.java
@@ -1,0 +1,26 @@
+package org.smartregister.chw.activity;
+
+import android.app.Activity;
+
+import org.smartregister.chw.R;
+import org.smartregister.chw.fragment.FamilyCallDialogFragment;
+import org.smartregister.chw.listener.OnClickFloatingMenu;
+import org.smartregister.chw.presenter.ChildProfilePresenter;
+
+public class ChildProfileActivityFlv {
+
+    public static OnClickFloatingMenu getOnClickFloatingMenu(final Activity activity, final ChildProfilePresenter presenter) {
+        return new OnClickFloatingMenu() {
+            @Override
+            public void onClickMenu(int viewId) {
+                switch (viewId) {
+                    case R.id.fab:
+                        FamilyCallDialogFragment.launchDialog(activity, presenter.getFamilyId());
+                        break;
+                    default:
+                        break;
+                }
+            }
+        };
+    }
+}
