+++ opensrp-chw/src/wcaro/java/org/smartregister/chw/repository/ChwRepositoryFlv.java
@@ -1,0 +1,124 @@
+package org.smartregister.chw.repository;
+
+import android.content.Context;
+import android.util.Log;
+
+import net.sqlcipher.database.SQLiteDatabase;
+
+import org.smartregister.domain.db.Column;
+import org.smartregister.immunization.repository.RecurringServiceRecordRepository;
+import org.smartregister.immunization.repository.VaccineRepository;
+import org.smartregister.immunization.util.IMDatabaseUtils;
+import org.smartregister.repository.AlertRepository;
+import org.smartregister.repository.EventClientRepository;
+
+public class ChwRepositoryFlv {
+
+    private static final String TAG = ChwRepositoryFlv.class.getCanonicalName();
+
+    public static void onUpgrade(Context context, SQLiteDatabase db, int oldVersion, int newVersion) {
+        Log.w(ChwRepository.class.getName(),
+                "Upgrading database from version " + oldVersion + " to "
+                        + newVersion + ", which will destroy all old data");
+        int upgradeTo = oldVersion + 1;
+        while (upgradeTo <= newVersion) {
+            switch (upgradeTo) {
+                case 2:
+                    upgradeToVersion2(context, db);
+                    break;
+                case 3:
+                    upgradeToVersion3(db);
+                    break;
+                case 4:
+                    upgradeToVersion4(db);
+                    break;
+                case 5:
+                    upgradeToVersion5(db);
+                    break;
+                case 7:
+                    upgradeToVersion7(db);
+                    break;
+                default:
+                    break;
+            }
+            upgradeTo++;
+        }
+    }
+
+
+    private static void upgradeToVersion2(Context context, SQLiteDatabase db) {
+        try {
+            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_EVENT_ID_COL);
+            db.execSQL(VaccineRepository.EVENT_ID_INDEX);
+            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_FORMSUBMISSION_ID_COL);
+            db.execSQL(VaccineRepository.FORMSUBMISSION_INDEX);
+
+            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_OUT_OF_AREA_COL);
+            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_OUT_OF_AREA_COL_INDEX);
+
+//            EventClientRepository.createTable(db, EventClientRepository.Table.path_reports, EventClientRepository.report_column.values());
+            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_HIA2_STATUS_COL);
+
+            IMDatabaseUtils.accessAssetsAndFillDataBaseForVaccineTypes(context, db);
+
+        } catch (Exception e) {
+            Log.e(TAG, "upgradeToVersion2 " + Log.getStackTraceString(e));
+        }
+
+    }
+
+    private static void upgradeToVersion3(SQLiteDatabase db) {
+        try {
+            Column[] columns = {EventClientRepository.event_column.formSubmissionId};
+            EventClientRepository.createIndex(db, EventClientRepository.Table.event, columns);
+
+            db.execSQL(VaccineRepository.ALTER_ADD_CREATED_AT_COLUMN);
+            VaccineRepository.migrateCreatedAt(db);
+
+            db.execSQL(RecurringServiceRecordRepository.ALTER_ADD_CREATED_AT_COLUMN);
+            RecurringServiceRecordRepository.migrateCreatedAt(db);
+        } catch (Exception e) {
+            Log.e(TAG, "upgradeToVersion3 " + Log.getStackTraceString(e));
+        }
+        try {
+            Column[] columns = {EventClientRepository.event_column.formSubmissionId};
+            EventClientRepository.createIndex(db, EventClientRepository.Table.event, columns);
+
+
+        } catch (Exception e) {
+            Log.e(TAG, "upgradeToVersion3 " + e.getMessage());
+        }
+    }
+
+    private static void upgradeToVersion4(SQLiteDatabase db) {
+        try {
+            db.execSQL(AlertRepository.ALTER_ADD_OFFLINE_COLUMN);
+            db.execSQL(AlertRepository.OFFLINE_INDEX);
+            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_TEAM_COL);
+            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_TEAM_ID_COL);
+            db.execSQL(RecurringServiceRecordRepository.UPDATE_TABLE_ADD_TEAM_COL);
+            db.execSQL(RecurringServiceRecordRepository.UPDATE_TABLE_ADD_TEAM_ID_COL);
+        } catch (Exception e) {
+            Log.e(TAG, "upgradeToVersion4 " + Log.getStackTraceString(e));
+        }
+
+    }
+
+    private static void upgradeToVersion5(SQLiteDatabase db) {
+        try {
+            db.execSQL(VaccineRepository.UPDATE_TABLE_ADD_CHILD_LOCATION_ID_COL);
+            db.execSQL(RecurringServiceRecordRepository.UPDATE_TABLE_ADD_CHILD_LOCATION_ID_COL);
+        } catch (Exception e) {
+            Log.e(TAG, "upgradeToVersion5 " + Log.getStackTraceString(e));
+        }
+    }
+
+    private static void upgradeToVersion7(SQLiteDatabase db) {
+        try {
+            db.execSQL(HomeVisitRepository.UPDATE_TABLE_ADD_VACCINE_NOT_GIVEN);
+            db.execSQL(HomeVisitRepository.UPDATE_TABLE_ADD_SERVICE_NOT_GIVEN);
+        } catch (Exception e) {
+            Log.e(TAG, "upgradeToVersion7 " + Log.getStackTraceString(e));
+        }
+    }
+}
